Chore: show blog page at '/' route
