+++ src/app/page.tsx
@@ -1,59 +1,27 @@
-"use client";
+import Image from "next/image";
+import { getFeaturedPosts, getSeries } from "@/utils/processPosts";
+import PostsCarousel from "./components/PostsCarousel";
+import Series from "./components/Series";
 
-import { useState } from "react";
-import Tags from "./components/tags";
-// TODO: constants로 이동
-const TAG = {
-  TAG1: "tag1",
-  TAG2: "tag2",
-  TAG3: "tag3",
-};
-const tags = [
-  {
-    type: TAG.TAG1,
-    value: "그로스에 대한 관심과 이해도, 경험이 있는",
-  },
-  {
-    type: TAG.TAG2,
-    value: "DX와 UX를 최적화하는 데 가치를 두는",
-  },
-  {
-    type: TAG.TAG3,
-    value: "꾸준히 함께 성장하는",
-  },
-];
-
-export interface TagData {
-  type: string;
-  value: string;
-}
-
-export default function Home() {
-  const [selectedTag, setSelectedTag] = useState(TAG.TAG1);
-  const handleTagSelect = (tag: string) => {
-    setSelectedTag(tag);
-  };
-
-  const translateSpace = {
-    [TAG.TAG1]: "translate-y-0",
-    [TAG.TAG2]: "-translate-y-13",
-    [TAG.TAG3]: "-translate-y-26",
-  };
+export default function index() {
+  const featuredPosts = getFeaturedPosts();
 
   return (
-    <div className="mt-20 flex justify-center h-screen">
-      <Tags.Sentence>
-        <Tags.List selectedTag={selectedTag} translateSpace={translateSpace}>
-          {tags.map((tag) => (
-            <Tags.Item
-              key={tag.type}
-              tag={tag}
-              isSelected={tag.type === selectedTag}
-              onClick={() => handleTagSelect(tag.type)}
-            />
+    <div className="flex flex-col items-center">
+      <section className="flex items-center h-40">
+        <Image src="/images/logo.svg" alt="logo" width={80} height={30} />
+        <p className="ml-10 whitespace-pre-wrap">
+          {`Why?에서 시작해 How를 찾는 과정을 좋아합니다. \n 그 고민과 성장의 과정을 꾸준히 기록하고자 합니다.`}
+        </p>
+      </section>
+      <div className="flex justify-center">
+        <main className="px-8 flex flex-col gap-y-40 max-w-lg md:max-w-5xl w-full">
+          <PostsCarousel title="추천 글" posts={featuredPosts} />
+          {getSeries().then((data) => (
+            <Series title="Series" posts={data} />
           ))}
-        </Tags.List>
-      </Tags.Sentence>
+        </main>
+      </div>
     </div>
   );
 }
